CURB-1449:
- actions refactor
