+++ libraries/engage/app-rating/actions/showModal.js
@@ -1,0 +1,70 @@
+import showModalAction from '@shopgate/pwa-common/actions/modal/showModal';
+import { getPlatform } from '@shopgate/pwa-common/selectors/client';
+import { historyPush } from '@shopgate/pwa-common/actions/router';
+import appConfig from '@shopgate/pwa-common/helpers/config';
+import { increaseRejectionCount, setLastPopupTimestamp } from '../action-creators/popup';
+import { generateReviewLink } from '../helpers';
+import { isMinTimeBetweenPopupsElapsed } from './minTime';
+
+const {
+  appRating: {
+    bundleId: bId,
+    rejectionLink,
+  },
+} = appConfig;
+
+/**
+ * shows the actual modal
+ * @param {Function} resetAction the reset action function
+ * @param {Function} increaseAction the function to increase the appropriate counter
+ * @param {boolean} mustShow if the modal must be shown
+ * @param {boolean} hasRepeats if the counters has repeats
+ * @return {(function(*, *): void)|*}
+ */
+export function showModal(resetAction, increaseAction, mustShow, hasRepeats) {
+  return (dispatch, getState) => {
+    if (!mustShow && hasRepeats) {
+      dispatch(increaseAction());
+    }
+
+    if (!(mustShow && hasRepeats)) {
+      return;
+    }
+
+    if (!isMinTimeBetweenPopupsElapsed()) {
+      return;
+    }
+
+    dispatch(resetAction());
+    dispatch(setLastPopupTimestamp());
+
+    dispatch(showModalAction({
+      confirm: 'appRating.yes',
+      dismiss: 'appRating.no',
+      title: 'appRating.title',
+      message: 'appRating.message',
+    })).then((confirmed) => {
+      // user touched yes and we
+      // redirect to store
+      if (confirmed) {
+        const platform = getPlatform(getState());
+        const link = generateReviewLink(bId[platform], platform);
+        if (!link) {
+          return;
+        }
+        historyPush({
+          pathname: link,
+        });
+        return;
+      }
+
+      // user doesn't want to rate
+      dispatch(increaseRejectionCount());
+      if (rejectionLink) {
+        historyPush({
+          pathname: rejectionLink,
+        });
+      }
+    });
+  };
+}
